+++ projects/angular-tag-cloud-module/src/lib/tag-cloud.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { TagCloudComponent } from './tag-cloud.component';
-
-@NgModule({
-  declarations: [TagCloudComponent],
-  exports: [TagCloudComponent],
-})
-export class TagCloudModule {}